Add Grammar-based Syntax Rules.
I've reverse engineered the grammar from the coffeescript suggestor script,
but I haven't yet implemented any of the security schemes.  I have nested the
keyword recognition withing regions so hopefully it should recognise the
correct keywords in the right places.  I've also enabled folding.

This syntax file is an extension to the YAML syntax file, so it should still
match most of the existing rules in the YAML syntax for constants and comments
and strings and what-not.

This is my first attempt at a syntax file.